Revert "Changes type of power_param from int to double."
This reverts commit cf20875f3fad4f32f22c8148bb4ebd312f7215bf.

+++ lib/WenoParams.h
@@ -9,7 +9,7 @@ public:
 		FD, JS, Z
 	} weno_version;
 	double epsilon;
-	double power_param;
+	int power_param;
 	double alpha_scaling;
 
 	void init(IniDocument& ini_doc);
